Refactored `storage_service.py` into a package structure to standardize with `social_auth`. The `storage` logic is now split into `base.py`, `cloudflare_r2.py`, and `factory.py`, matching the architecture of other services. Updated imports across the codebase and ran tests to verify everything passes.
