+++ backend/apps/core/services/storage/__init__.py
@@ -1,0 +1,10 @@
+from .base import StorageService
+from .cloudflare_r2 import CloudflareR2StorageService
+from .factory import get_storage_service
+
+
+__all__ = [
+    "CloudflareR2StorageService",
+    "StorageService",
+    "get_storage_service",
+]
