+++ backend/apps/core/tests/test_security_audit.py
@@ -100,7 +100,9 @@ class TestSecurityAudit:
             "apps/users/services/token_service.py",
             "apps/users/services/google_auth_service.py",
             "apps/tenants/services/tenant_service.py",
-            "apps/core/services/storage_service.py",
+            "apps/core/services/storage/base.py",
+            "apps/core/services/storage/cloudflare_r2.py",
+            "apps/core/services/storage/factory.py",
             "apps/core/services/social_auth/base.py",
             "apps/core/services/social_auth/google_provider.py",
             "apps/scheduler/services/templates.py",
